team_p04_watanabe/src/filtering/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "filter-solvable-question"
version = "0.1.0"
description = "Filter solvable questions using LLM evaluation"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Project Author", email = "author@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "datasets>=2.16.0",
    "vllm>=0.3.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "filelock>=3.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.scripts]
evaluate = "filter_solvable_question.cli.main:main"
aggregate = "filter_solvable_question.cli.aggregate:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[dependency-groups]
dev = [
    "ipython>=9.4.0",
]
